make it all compile, cleanup
